[RF][Start][TeamPage]Refactoring TeamPage, AbcTest, LanguageMathematicaTest, LanguageJava3Test, TeamTest; deleting GalaKonTest (#530)
* Refactoring TeamPage, AbcTest, LanguageMathematicaTest, LanguageJava3Test, TeamTest; deleting GalaKonTest

* testBrowseLanguagesSubmenuText() is deleted because the similar test already exist in main branch in the project
